Fix final compilation errors
- Added CurrentLevel property to GameManager
- Added GetPlayerRank and GetPlayerId to ILeaderboardService
- Fixed long to int conversions in AchievementService.UpdateFromStats
- Fixed long to int conversions in DailyChallengeService.OnCurrencyChanged

+++ Assets/Scripts/Systems/DailyChallengeService.cs
@@ -379,11 +379,11 @@ namespace Rumbax.Systems
             {
                 if (evt.Delta > 0)
                 {
-                    UpdateChallengeProgress(ChallengeType.EarnCoins, evt.Delta);
+                    UpdateChallengeProgress(ChallengeType.EarnCoins, (int)evt.Delta);
                 }
                 else
                 {
-                    UpdateChallengeProgress(ChallengeType.SpendCoins, -evt.Delta);
+                    UpdateChallengeProgress(ChallengeType.SpendCoins, (int)(-evt.Delta));
                 }
             }
         }
